# Group code objects by kernel name in perf report summary