@umangyadav 2022-06-02
> I cannot get the models to run in driver perf.
> 
> ```
> root@557fcf91b2a1:/code/2_AMDMIGraphX/docker_build# ./bin/driver perf bertsquad-10.onnx --fill1 input_ids --fill1 segment_ids --input-dim @input_ids 1 384 --fp16
> Compiling ...
> Reading: bertsquad-10.onnx
> terminate called after throwing an instance of 'migraphx::version_1::exception'
>   what():  /code/2_AMDMIGraphX/src/program.cpp:171: compile: Dangling reference in module main from instruction 998
> Aborted (core dumped)
> ```
@CharlieL7 
Maybe same as this issue #1230