@causten 2022-06-02
#1230  was fixed with https://github.com/ROCmSoftwarePlatform/AMDMIGraphX/pull/1233.  And that's merged