@causten 2022-06-02
I too am seeing a Dangling reference problem, but it is in the develop branch, not this PR.  I ran 
/opt/rocm/bin/migraphx-driver perf --model resnet50 
with success.